consolidate Factor components
