+++ client/src/modules/DecisionLab/FactorCard/FactorCard.tsx
@@ -1,1 +1,5 @@
-export const FactorCard = () => <div>Factor Card</div>;
+import { Card } from "@/client/components/Card/Card";
+
+export const FactorCard = ({ factor }: { factor: Factor }) => (
+  <Card title={factor.name} body={factor.description + " " + factor.type} />
+);
