+++ types/decision.d.ts
@@ -8,5 +8,13 @@ interface Decision {
   score: number;
 }
 
+interface Factor {
+  id: string;
+  name: string;
+  description: string;
+  type: "pro" | "con";
+  score: number;
+}
+
 type CreateDecisionDto = Pick<Decision, "title">;
 type UpdateDecisionDto = Omit<Decision, "id">;
